+++ src/main/java/com/aaronpmaus/jMath/executables/FindMaxClique.java
@@ -1,5 +1,9 @@
 package com.aaronpmaus.jMath.executables;
+
 import com.aaronpmaus.jMath.graph.*;
+import com.aaronpmaus.jMath.io.GraphIO;
+import com.aaronpmaus.jMath.io.CommandLineParser;
+
 import java.util.ArrayList;
 import java.util.Set;
 import java.util.Date;