+++ src/main/java/com/aaronpmaus/jMath/graph/IncMaxCliqueAdapter.java
@@ -1,4 +1,5 @@
 package com.aaronpmaus.jMath.graph;
+
 import java.io.FileNotFoundException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -10,6 +11,8 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Date;
+
+import com.aaronpmaus.jMath.io.GraphIO;
 
 /**
 * A class that uses IncMaxCliqueSolver from Combining MaxSAT Reasoning and Incremental Upper Bound