+++ CHANGELOG.md
@@ -19,7 +19,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 [None]
 
-## [0.13.0]
+## [0.13.1]
+### Added
+- Package io with:
+  - CommandLineParser (new class)
+  - GraphIO (moved from graph package)
+
+## [0.13.0] - 2018-04-14
 ### Added
 - Add Depth First Search to Graph and junit test to UndirectedGraph tester
 ### Fixed